+++ Plugins/TableProPluginKit/ConnectionField.swift
@@ -7,6 +7,10 @@ public enum FieldSection: String, Codable, Sendable {
     case connection
 }
 
+public enum DynamicFieldOptions: String, Codable, Sendable {
+    case awsProfiles
+}
+
 public struct FieldVisibilityRule: Codable, Sendable, Equatable {
     public let fieldId: String
     public let values: [String]
@@ -92,6 +96,7 @@ public struct ConnectionField: Codable, Sendable {
     public let section: FieldSection
     public let hidesPassword: Bool
     public let visibleWhen: FieldVisibilityRule?
+    public var dynamicOptions: DynamicFieldOptions?
 
     /// Backward-compatible convenience: true when fieldType is .secure
     public var isSecure: Bool {
@@ -122,6 +127,12 @@ public struct ConnectionField: Codable, Sendable {
         self.visibleWhen = visibleWhen
     }
 
+    public func withDynamicOptions(_ options: DynamicFieldOptions?) -> ConnectionField {
+        var copy = self
+        copy.dynamicOptions = options
+        return copy
+    }
+
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         id = try container.decode(String.self, forKey: .id)
@@ -133,9 +144,11 @@ public struct ConnectionField: Codable, Sendable {
         section = try container.decodeIfPresent(FieldSection.self, forKey: .section) ?? .advanced
         hidesPassword = try container.decodeIfPresent(Bool.self, forKey: .hidesPassword) ?? false
         visibleWhen = try container.decodeIfPresent(FieldVisibilityRule.self, forKey: .visibleWhen)
+        dynamicOptions = try container.decodeIfPresent(DynamicFieldOptions.self, forKey: .dynamicOptions)
     }
 
     private enum CodingKeys: String, CodingKey {
         case id, label, placeholder, isRequired, defaultValue, fieldType, section, hidesPassword, visibleWhen
+        case dynamicOptions
     }
 }